requirements.txt
altair==4.2.0
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.0.0
fastapi==0.85.1
fastjsonschema==2.16.1
fonttools==4.37.2
gitdb==4.0.9
GitPython==3.1.27
h11==0.14.0
idna==3.4
imbalanced-learn==0.9.1
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jose==1.0.0
json5==0.9.10
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.1
kiwisolver==1.4.4
logger==1.4
lottie==0.7.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pydeck==0.8.0b3
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.0
requests==2.28.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.0
semver==2.13.0
Send2Trash==1.8.0
setuptools==65.4.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.5.0
starlette==0.20.4
streamlit==1.13.0
streamlit-lottie==0.0.3
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.4.0
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
uvicorn==0.19.0
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
wheel==0.37.1
xgboost==1.6.2
zipp==3.8.1
